requirements.txt
python == 3.6.8
gensim~=4.0.1
matplotlib~=3.2.2
nltk~=3.5
numpy~=1.18.5
pandas~=1.0.5
scipy~=1.5.0
seaborn~=0.10.1
scikit-image~=0.18.2
torch~=1.9.0
torchnet == 0.0.4
torchvision~=0.10.0
umap-learn == 0.1.1

tqdm~=4.46.1
Pillow~=7.1.2
visdom~=0.1.8.9
six~=1.15.0
opencv-python~=4.3.0.36
reprint~=0.6.0
torchinfo~=1.5.3
requests~=2.24.0
scikit-learn~=0.23.1